+++ pilad/conn.go
@@ -9,7 +9,6 @@ import (
 	"github.com/fern4lvarez/piladb/config"
 	"github.com/fern4lvarez/piladb/pila"
 	"github.com/fern4lvarez/piladb/pkg/uuid"
-	"github.com/fern4lvarez/piladb/pkg/version"
 
 	"github.com/gorilla/mux"
 )
@@ -33,17 +32,24 @@ func NewConn() *Conn {
 	conn := &Conn{}
 	conn.Pila = pila.NewPila()
 	conn.Config = config.NewConfig()
-	conn.Status = NewStatus(version.Version(version.VERSION), time.Now().UTC(), MemStats())
+	conn.Status = NewStatus(v(), time.Now().UTC(), MemStats())
 	return conn
 }
 
 // Connection Handlers
 
-// rootHandler redirects to the pilad documentation site hosted on Github.
+// rootHandler shows information about piladb.
 func (c *Conn) rootHandler(w http.ResponseWriter, r *http.Request) {
-	redirAddress := fmt.Sprintf("https://raw.githubusercontent.com/fern4lvarez/piladb/%s/pilad/README.md", version.CommitHash())
-	log.Println(r.Method, r.URL, http.StatusMovedPermanently, "Moved to", redirAddress)
-	http.Redirect(w, r, redirAddress, http.StatusMovedPermanently)
+	var links = []byte(`{"thank you":"for using piladb","www":"https://www.piladb.org","code":"https://github.com/fern4lvarez/piladb","docs":"https://docs.piladb.org"}`)
+	w.Header().Set("Content-Type", "application/json")
+	log.Println(r.Method, r.URL, http.StatusOK)
+	w.Write(links)
+}
+
+// pingHandler writes pong.
+func (c *Conn) pingHandler(w http.ResponseWriter, r *http.Request) {
+	log.Println(r.Method, r.URL, http.StatusOK)
+	w.Write([]byte("pong"))
 }
 
 // statusHandler writes the piladb status into the response.