+++ CHANGELOG.md
@@ -5,9 +5,20 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.1.2] - 2017-03-05
+
 ### Added
 - pilad: Add Go Version to Status
+- pilad: Add `/_ping` endpoint
 - godoc: Extend packages documentation
+- vendor: Update dependencies
+- dev: Add pila.sh utilities to Docker image
+
+### Changed
+- pilad: Show links of interest in `/` endpoint.
+
+### Fixed
+- Fix decoding bug when pushing a malformed payload
 
 ## [0.1.1] - 2017-02-20
 
@@ -24,6 +35,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - First release!
 
-[Unreleased]: https://github.com/fern4lvarez/piladb/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/fern4lvarez/piladb/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/fern4lvarez/piladb/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/fern4lvarez/piladb/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/fern4lvarez/piladb/compare/dda6b656cbd635dab8e9fc6c254a46f01e4e43ca...v0.1.0